debug build error: add suspense to use search params
