+++ ui/components/common/referrer-handler.tsx
@@ -1,0 +1,20 @@
+"use client";
+
+import { useSearchParams } from "next/navigation";
+import { useEffect } from "react";
+
+const ReferrerHandler = ({ setReferrer }: { setReferrer: any }) => {
+  const searchParams = useSearchParams();
+  const referredByID = searchParams.get("referralid");
+  console.log("referredByID", referredByID);
+
+  useEffect(() => {
+    if (referredByID) {
+      setReferrer(referredByID);
+    }
+  }, [referredByID]);
+
+  return null;
+};
+
+export default ReferrerHandler;
